Stop when slowdown!
We wanna ensure the safety,but also don't slowdown it's efficiency.
So we change the condition when should we stop the extension and contraction move.
Original condition: When it stop move: count == count1?
New condition: When the speed of extension or contraction slowdown,which means that it may encounter obstacles.